+++ offchain/crates/validator-debt/src/worker.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, str::FromStr};
+use std::{collections::HashMap, str::FromStr, sync::Arc};
 
 use anyhow::{Context, Result, bail, ensure};
 use doublezero_solana_client_tools::{
@@ -23,7 +23,9 @@ use doublezero_solana_sdk::{
     },
     try_build_instruction,
 };
+use futures::{StreamExt, TryStreamExt, stream};
 use leaky_bucket::RateLimiter;
+use reqwest::Client;
 use serde::Serialize;
 use slack_notifier;
 use solana_client::nonblocking::rpc_client::RpcClient;
@@ -386,26 +388,67 @@ pub async fn calculate_distribution(
     Ok(write_summary)
 }
 
-pub async fn pay_solana_validator_debt(
+pub async fn pay_all_solana_validator_debt(
     wallet: Wallet,
     dz_ledger: DoubleZeroLedgerConnection,
-    dz_epoch: u64,
-) -> Result<DebtCollectionResults> {
+) -> Result<()> {
     let (_, config) = try_fetch_config(&wallet.connection).await?;
+    let dz_epoch_range = Vec::from_iter(
+        GENESIS_DZ_EPOCH_MAINNET_BETA..(config.last_completed_epoch().unwrap().value()),
+    );
 
+    let tasks: Vec<DebtCollectionResults> = stream::iter(dz_epoch_range)
+        .map(|dz_epoch| {
+            let wallet_ref = &wallet;
+            let ledger_ref = &dz_ledger;
+            let config_ref = &config;
+
+            async move {
+                let result =
+                    pay_solana_validator_debt(wallet_ref, ledger_ref, dz_epoch, config_ref).await?;
+                tracing::info!("Finished debt collection for epoch {dz_epoch}");
+                Ok::<_, anyhow::Error>(result)
+            }
+        })
+        .buffer_unordered(2)
+        .try_collect()
+        .await?;
+
+    let client = reqwest::Client::new();
+
+    post_debt_collection_summary_to_slack(&tasks, &client).await?;
+    post_debt_collections_to_slack(&tasks, false, &client).await?;
+
+    Ok(())
+}
+
+pub async fn pay_solana_validator_debt(
+    wallet: &Wallet,
+    dz_ledger: &DoubleZeroLedgerConnection,
+    dz_epoch: u64,
+    config: &ProgramConfig,
+) -> Result<DebtCollectionResults> {
     let (_, computed_debt) = ledger::try_fetch_debt_record(
-        &dz_ledger,
+        dz_ledger,
         &config.debt_accountant_key,
         dz_epoch,
         dz_ledger.commitment(),
     )
     .await?;
-    try_initialize_missing_deposit_accounts(&wallet, &computed_debt).await?;
 
-    let transaction = Transaction::new(wallet.signer, wallet.dry_run, false);
+    let (distribution_key, _) = Distribution::find_address(DoubleZeroEpoch::new(dz_epoch));
+    let distribution = wallet
+        .connection
+        .try_fetch_zero_copy_data::<Distribution>(&distribution_key)
+        .await?;
+
+    try_initialize_missing_deposit_accounts(wallet, &computed_debt).await?;
+
+    let arc_signer = Arc::new(wallet.signer.insecure_clone());
+    let transaction = Transaction::new(arc_signer, wallet.dry_run, false);
 
     let tx_results = transaction
-        .pay_solana_validator_debt(&wallet.connection, computed_debt, dz_epoch)
+        .pay_solana_validator_debt(&wallet.connection, computed_debt, dz_epoch, &distribution)
         .await?;
     Ok(tx_results)
 }
@@ -575,6 +618,99 @@ pub async fn try_initialize_distribution(
     Ok(())
 }
 
+pub async fn post_debt_collection_summary_to_slack(
+    debt_collection_results: &[DebtCollectionResults],
+    client: &Client,
+) -> Result<()> {
+    let total_paid: u64 = debt_collection_results.iter().map(|tp| tp.total_paid).sum();
+    let total_debt: u64 = debt_collection_results.iter().map(|td| td.total_debt).sum();
+    let insufficient_funds_count: usize = debt_collection_results
+        .iter()
+        .map(|ifc| ifc.insufficient_funds_count)
+        .sum();
+
+    let header = "Total Debt Collection";
+    let table_header = vec![
+        "Total Paid".to_string(),
+        "Total Debt".to_string(),
+        "Total Outstanding".to_string(),
+        "Total Insufficient Funds Count".to_string(),
+    ];
+    let table_values = vec![
+        format!("{:.9} SOL", total_paid as f64 * 1e-9),
+        format!("{:.9} SOL", total_debt as f64 * 1e-9),
+        format!("{:.9} SOL", (total_debt - total_paid) as f64 * 1e-9),
+        insufficient_funds_count.to_string(),
+    ];
+    slack_notifier::validator_debt::post_to_slack(None, client, header, table_header, table_values)
+        .await?;
+    Ok(())
+}
+
+pub async fn post_debt_collections_to_slack(
+    debt_collection_results: &Vec<DebtCollectionResults>,
+    dry_run: bool,
+    client: &Client,
+) -> Result<()> {
+    let header = if dry_run {
+        "DRY RUN Debt Collected DRY RUN"
+    } else {
+        "Debt Collected"
+    };
+
+    let table_header = vec![
+        "DoubleZero Epoch".to_string(),
+        "Total Paid".to_string(),
+        "Outstanding Debt".to_string(),
+        "Total Debt".to_string(),
+        "Percentage Paid".to_string(),
+        "Total Attempted Transactions".to_string(),
+        "Successful Transactions".to_string(),
+        "Insufficient Funds".to_string(),
+        "Already Paid".to_string(),
+    ];
+
+    let mut table_values: Vec<Vec<String>> = Vec::with_capacity(debt_collection_results.len());
+
+    for dcr in debt_collection_results {
+        let total_attempted_transactions_count: u64 = dcr.total_validators as u64;
+
+        // skip overlapping DZ epochs since we don't collect fees
+        if total_attempted_transactions_count == 0 {
+            continue;
+        };
+        let successful_transactions_count: u64 = dcr.successful_transactions_count as u64;
+        let already_paid_count: u64 = dcr.already_paid_count as u64;
+
+        let percentage_paid = (already_paid_count + successful_transactions_count) as f64
+            / total_attempted_transactions_count as f64;
+
+        let row_values = vec![
+            dcr.dz_epoch.to_string(),
+            format!("{:.9} SOL", dcr.total_paid as f64 * 1e-9),
+            format!("{:.9} SOL", (dcr.total_debt - dcr.total_paid) as f64 * 1e-9),
+            format!("{:.9} SOL", dcr.total_debt as f64 * 1e-9),
+            format!("{:.2}%", percentage_paid * 100.0),
+            total_attempted_transactions_count.to_string(),
+            successful_transactions_count.to_string(),
+            dcr.insufficient_funds_count.to_string(),
+            already_paid_count.to_string(),
+        ];
+
+        table_values.push(row_values);
+    }
+
+    slack_notifier::validator_debt::post_debt_collections_to_slack(
+        client,
+        header,
+        table_header,
+        table_values,
+    )
+    .await?;
+
+    Ok(())
+}
+
 pub async fn post_debt_collection_to_slack(
     debt_collection_results: DebtCollectionResults,
     dry_run: bool,
@@ -590,6 +726,7 @@ pub async fn post_debt_collection_to_slack(
     let table_header = vec![
         "DoubleZero Epoch".to_string(),
         "Total Paid".to_string(),
+        "Outstanding Debt".to_string(),
         "Total Debt".to_string(),
         "Percentage Paid".to_string(),
         "Total Attempted Transactions".to_string(),
@@ -598,9 +735,10 @@ pub async fn post_debt_collection_to_slack(
         "Already Paid".to_string(),
     ];
 
-    let total_attempted_transactions_count = debt_collection_results.collection_results.len();
-    let successful_transactions_count = debt_collection_results.successful_transactions.len();
-    let already_paid_count = debt_collection_results.already_paid.len();
+    let total_attempted_transactions_count: u64 = debt_collection_results.total_validators as u64;
+    let successful_transactions_count: u64 =
+        debt_collection_results.successful_transactions_count as u64;
+    let already_paid_count: u64 = debt_collection_results.already_paid_count as u64;
 
     let percentage_paid: f64 = if total_attempted_transactions_count == 0 {
         0.0
@@ -609,32 +747,24 @@ pub async fn post_debt_collection_to_slack(
             / total_attempted_transactions_count as f64
     };
 
-    // the total amount paid for an epoch is `total_collected_this_run` + `already_paid`
-    let already_paid_total: u64 = debt_collection_results
-        .already_paid
-        .iter()
-        .map(|ap| ap.amount)
-        .sum();
-    let total_collected_this_run: u64 = debt_collection_results
-        .successful_transactions
-        .iter()
-        .map(|ap| ap.amount)
-        .sum();
-
-    let total_paid = already_paid_total + total_collected_this_run;
-    let total_debt: u64 = debt_collection_results
-        .collection_results
-        .iter()
-        .map(|cr| cr.amount)
-        .sum();
     let table_values = vec![
         debt_collection_results.dz_epoch.to_string(),
-        total_paid.to_string(),
-        total_debt.to_string(),
+        format!(
+            "{:.9} SOL",
+            debt_collection_results.total_paid as f64 * 1e-9
+        ),
+        format!(
+            "{:.9} SOL",
+            (debt_collection_results.total_debt - debt_collection_results.total_paid) as f64 * 1e-9
+        ),
+        format!(
+            "{:.9} SOL",
+            debt_collection_results.total_debt as f64 * 1e-9
+        ),
         format!("{:.2}%", percentage_paid * 100.0),
         total_attempted_transactions_count.to_string(),
         successful_transactions_count.to_string(),
-        debt_collection_results.insufficient_funds.len().to_string(),
+        debt_collection_results.insufficient_funds_count.to_string(),
         already_paid_count.to_string(),
     ];