+++ offchain/scheduler/lib/scheduler/scheduler_doublezero.ex
@@ -7,7 +7,11 @@ defmodule Scheduler.DoubleZero do
 
   def initialize_tracing_subscriber, do: :erlang.nif_error(:nif_not_loaded)
 
-  def pay_debt(_dz_epoch, _ledger_rpc, _solana_rpc), do: :erlang.nif_error(:nif_not_loaded)
+  def collect_all_debt(_ledger_rpc, _solana_rpc), do: :erlang.nif_error(:nif_not_loaded)
+
+  def collect_epoch_debt(_dz_epoch, _ledger_rpc, _solana_rpc),
+    do: :erlang.nif_error(:nif_not_loaded)
+
   def initialize_distribution(_solana_rpc), do: :erlang.nif_error(:nif_not_loaded)
 
   def calculate_distribution(_dz_epoch, _ledger_rpc, _solana_rpc, _post_to_slack),
@@ -17,7 +21,4 @@ defmodule Scheduler.DoubleZero do
     do: :erlang.nif_error(:nif_not_loaded)
 
   def current_dz_epoch(_ledger_rpc), do: :erlang.nif_error(:nif_not_loaded)
-
-  def post_debt_summary(_insufficient_funds_count, _total_debt, _total_paid),
-    do: :erlang.nif_error(:nif_not_loaded)
 end