+++ offchain/scheduler/lib/scheduler/validator_debt/debt_collection.ex
@@ -1,5 +1,6 @@
 defmodule Scheduler.ValidatorDebt.DebtCollection do
   defstruct total_paid: 0,
+            already_paid: 0,
             total_debt: 0,
             total_validators: 0,
             insufficient_funds_count: 0