+++ offchain/scheduler/native/scheduler_doublezero/src/lib.rs
@@ -1,8 +1,11 @@
+use std::sync::Arc;
+
 use anyhow::Result;
 use doublezero_solana_client_tools::{
     payer::{Wallet, try_load_keypair},
     rpc::{DoubleZeroLedgerConnection, SolanaConnection},
 };
+use doublezero_solana_sdk::revenue_distribution::fetch::try_fetch_config;
 use doublezero_solana_validator_debt::{
     rpc::SolanaValidatorDebtConnectionOptions,
     solana_debt_calculator::SolanaDebtCalculator,
@@ -16,8 +19,11 @@ use tracing_subscriber::{EnvFilter, layer::SubscriberExt, util::SubscriberInitEx
 #[derive(NifStruct)]
 #[module = "Scheduler.ValidatorDebt.DebtCollection"]
 pub struct DebtCollection {
+    pub dz_epoch: u64,
     pub total_paid: u64,
     pub total_debt: u64,
+    pub already_paid: u64,
+    pub outstanding_debt: u64,
     pub total_validators: usize,
     pub insufficient_funds_count: usize,
 }
@@ -46,46 +52,8 @@ pub fn initialize_tracing_subscriber() -> Result<(), NifError> {
     Ok(())
 }
 
-#[rustler::nif]
-pub fn post_debt_summary(
-    insufficient_funds_count: usize,
-    total_debt: u64,
-    total_paid: u64,
-) -> Result<(), NifError> {
-    Runtime::new()
-        .map_err(display_to_nif_error)?
-        .block_on(async {
-            let client = reqwest::Client::new();
-
-            let header = "Total Debt Collection";
-            let table_header = vec![
-                "Total Paid".to_string(),
-                "Total Debt".to_string(),
-                "Total Insufficient Funds Count".to_string(),
-            ];
-            let table_values = vec![
-                total_paid.to_string(),
-                total_debt.to_string(),
-                insufficient_funds_count.to_string(),
-            ];
-            slack_notifier::validator_debt::post_to_slack(
-                None,
-                &client,
-                header,
-                table_header,
-                table_values,
-            )
-            .await?;
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .map_err(display_to_nif_error)?;
-
-    Ok(())
-}
-
 #[rustler::nif(schedule = "DirtyIo")]
-pub fn pay_debt(
+pub fn collect_epoch_debt(
     dz_epoch: u64,
     ledger_rpc: String,
     solana_rpc: String,
@@ -94,41 +62,27 @@ pub fn pay_debt(
     let tx_results = Runtime::new()
         .map_err(display_to_nif_error)?
         .block_on(async {
-            let ledger_rpc_client = DoubleZeroLedgerConnection::new(ledger_rpc);
-
             let wallet = try_initialize_wallet(solana_rpc)?;
+            let ledger_rpc_client = DoubleZeroLedgerConnection::new(ledger_rpc);
+            let (_, config) = try_fetch_config(&wallet.connection).await?;
 
             let tx_results =
-                worker::pay_solana_validator_debt(wallet, ledger_rpc_client, dz_epoch).await?;
+                worker::pay_solana_validator_debt(&wallet, &ledger_rpc_client, dz_epoch, &config)
+                    .await?;
 
             worker::post_debt_collection_to_slack(tx_results.clone(), false, None).await?;
 
             Ok::<DebtCollectionResults, anyhow::Error>(tx_results)
         })
         .map_err(display_to_nif_error)?;
-
-    let total_validators = tx_results.collection_results.len();
-
-    let total_debt: u64 = tx_results
-        .collection_results
-        .iter()
-        .map(|tx| tx.amount)
-        .sum();
-
-    let insufficient_funds_count = tx_results.insufficient_funds.len();
-
-    let already_paid: u64 = tx_results.already_paid.iter().map(|tx| tx.amount).sum();
-    let successful_transactions: u64 = tx_results
-        .successful_transactions
-        .iter()
-        .map(|tx| tx.amount)
-        .sum();
-    let total_paid = already_paid + successful_transactions;
     let debt_collection = DebtCollection {
-        total_debt,
-        total_paid,
-        total_validators,
-        insufficient_funds_count,
+        dz_epoch: tx_results.dz_epoch,
+        already_paid: tx_results.already_paid,
+        total_debt: tx_results.total_debt,
+        total_paid: tx_results.total_paid,
+        outstanding_debt: (tx_results.total_debt - tx_results.total_paid),
+        total_validators: tx_results.total_validators,
+        insufficient_funds_count: tx_results.insufficient_funds_count,
     };
 
     Ok(debt_collection)
@@ -151,6 +105,28 @@ pub fn initialize_distribution(solana_rpc: String) -> Result<(), NifError> {
         })
         .map_err(display_to_nif_error)?;
 
+    Ok(())
+}
+
+#[rustler::nif(schedule = "DirtyIo")]
+pub fn collect_all_debt(ledger_rpc: String, solana_rpc: String) -> Result<(), NifError> {
+    let ledger_rpc_client = DoubleZeroLedgerConnection::new(ledger_rpc);
+
+    Runtime::new()
+        .map_err(display_to_nif_error)?
+        .block_on(async {
+            let wallet = Wallet {
+                connection: SolanaConnection::new(solana_rpc),
+                signer: try_load_keypair(None)?,
+                compute_unit_price_ix: None,
+                verbose: false,
+                fee_payer: None,
+                dry_run: false,
+            };
+
+            worker::pay_all_solana_validator_debt(wallet, ledger_rpc_client).await
+        })
+        .map_err(display_to_nif_error)?;
     Ok(())
 }
 
@@ -185,8 +161,9 @@ pub fn calculate_distribution(
             };
             let solana_debt_calculator: SolanaDebtCalculator =
                 SolanaDebtCalculator::try_from(connection_options)?;
-
-            let transaction = Transaction::new(try_load_keypair(None)?, false, false);
+            let keypair = try_load_keypair(None)?;
+            let arc_keypair = Arc::new(keypair);
+            let transaction = Transaction::new(arc_keypair, false, false);
 
             let write_summary = worker::calculate_distribution(
                 &solana_debt_calculator,
@@ -231,7 +208,9 @@ pub fn finalize_distribution(
             let solana_debt_calculator: SolanaDebtCalculator =
                 SolanaDebtCalculator::try_from(connection_options)?;
 
-            let transaction = Transaction::new(try_load_keypair(None)?, false, false);
+            let keypair = try_load_keypair(None)?;
+            let arc_keypair = Arc::new(keypair);
+            let transaction = Transaction::new(arc_keypair, false, false);
 
             worker::finalize_distribution(&solana_debt_calculator, transaction, dz_epoch).await?;